[REF] Rename data to data_itau
