+++ tests/test_arquivo.py
@@ -10,7 +10,7 @@ import codecs
 from cnab240 import errors
 from cnab240.bancos import itau
 from cnab240.tipos import Arquivo
-from tests.data import get_itau_data_from_dict, get_itau_file_remessa, \
+from tests.data_itau import get_itau_data_from_dict, get_itau_file_remessa, \
     ARQS_DIRPATH
