+++ tests/test_evento.py
@@ -6,7 +6,6 @@ except ImportError:
 
 from cnab240.bancos import itau
 from cnab240.tipos import Evento
-from tests.data import get_itau_data_from_file
 
 
 class TestEvento(unittest.TestCase):
