+++ tests/test_registro.py
@@ -7,7 +7,7 @@ except ImportError:
 from decimal import Decimal
 from cnab240 import errors
 from cnab240.bancos import itau
-from tests.data import get_itau_data_from_file
+from tests.data_itau import get_itau_data_from_file
 
 
 class TestRegistro(unittest.TestCase):
